+++ src/backend/opencl/traits.hpp
@@ -10,6 +10,7 @@
 #pragma once
 
 #include <defines.hpp>
+#include <types.hpp>
 #include <af/traits.hpp>
 #include <string>
 #include <sstream>
